requirements.txt
Flask==1.1.2
Werkzeug==0.15.1
tabula_py==2.1.1
pandas==0.25.3
SQLAlchemy==1.3.18
mysql_connector_repackaged==0.3.1
PyPDF2==1.26.0
python-dotenv==0.14.0
tabula==1.0.5
